requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart==0.0.20

# Auth
authlib==1.3.0
python-jose[cryptography]==3.3.0
httpx==0.28.1

# Database
aiosqlite==0.19.0

# Session
itsdangerous==2.1.2

# AI
google-genai==1.21.0

# MCP (Model Context Protocol)
mcp[http]>=1.9.3

# Bright Data SDK
brightdata-sdk>=0.1.0

# Blockchain
web3==6.15.0
eth-account==0.11.0

# Utils
python-dotenv==1.1.0
pydantic>=2.8.0
pydantic-settings>=2.1.0
tenacity>=8.2.3
